Add empty model-new and extract some components
