+++ app/assets/javascripts/pages/boards/show/new_issue_form.coffee
@@ -1,3 +1,4 @@
+# TODO: Remove after refactoring
 class @NewIssueForm extends View
   initialize: ->
     @$form_content = @$('.create-issue')
