+++ app/assets/stylesheets/pages/boards/show.sass
@@ -199,6 +199,7 @@ $image-path: 'pages/p-boards-show'
           .add-user
             display: none
 
+  // TODO: Remove this
   .create-issue
     background: rgba(90,90,90,0.3)
     display: none
