+++ app/assets/stylesheets/pages/issues/new.sass
@@ -1,3 +1,4 @@
+// TODO: Remove this
 .p-issues-new
   form
     margin-top: 20px
