+++ app/views/boards/show.html.slim
@@ -12,6 +12,7 @@
   ]
 
   = render partial: 'issues/modal'
+  = render partial: 'issues/modal_new'
 
   .board-columns
     - @board_bag.columns.each do |column|
