+++ app/views/columns/_menu.html.slim
@@ -8,5 +8,6 @@
 
   - if can?(:update, @board) && k(column).first?
     .new-issue.ajax-link Add an issue
-    .create-issue
+    // TODO: Remove it
+    /.create-issue
       = render 'issues/new', board_bag: @board_bag, issue: @board_bag.build_issue_new
