+++ app/views/issues/_modal_new.html.slim
@@ -1,0 +1,9 @@
+// TODO: Add labels json
+.issue-modal-container [
+    data-labels="#{[
+      { id: 'bug', name: 'bug', color: '#fff', backgroundColor: '#d93f0b', checked: false},
+      { id: 'duplicate', name: 'duplicate', color: '#000', backgroundColor: '#ccc', checked: false }
+    ].to_json}"
+  ]
+  .issue-modal-new#issue-modal-new
+
